Hero image successfully aligned
The extra div of the image removed, the flex-basis of each bottom section has been setted to 50%, cause we want each sextion to have as a base of its width the half of its parent's.
The height and the width of the image are at 100%, cause we want the image to fill the entire div container.